requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.6
py2neo==2021.2.3
PyGithub==1.55
Pillow==9.4.0
Jinja2==3.1.2
pdflatex==0.1.3
python-dotenv==1.0.0
requests==2.31.0
# For image processing
numpy==1.24.3
# For security
bcrypt==4.0.1
